+++ examples/api-authentication/client/scripts/controllers.js
@@ -6,47 +6,70 @@ myApp.controller('MenuCtrl', function ($scope, $location) {
   };
 });
 
+myApp.controller('MsgCtrl', function ($scope, auth) {
+  $scope.message = 'loading...';
+  auth.loaded.then(function () {
+    $scope.message = '';
+  });
+});
+
 myApp.controller('RootCtrl', function (auth, $scope, $location, $http, API_ENDPOINT) {
+  $scope.$parent.message = 'Welcome ' + auth.profile.name + '!';
   $scope.auth = auth;
 
   $scope.sendProtectedMessage = function () {
     $http({method: 'GET', url: API_ENDPOINT})
       .success(function (data, status, headers, config) {
-        $scope.result = 'Protected data was: ' + data;
+        $scope.$parent.message = 'Protected data was: ' + data;
       });
   };
 });
 
-myApp.controller('LoginCtrl', function (auth, $scope) {
+myApp.controller('LoginCtrl', function (auth, $scope, $location) {
   $scope.user = '';
   $scope.pass = '';
-      
+
+  function onLoginSuccess() {
+    $scope.$parent.message = '';
+    $location.path('/');
+  }
+
+  function onLoginFailed() {
+    $scope.$parent.message = 'invalid credentials';
+  }
+
   $scope.submit = function () {
+    $scope.$parent.message = 'loading...';
+    $scope.loading = true;
+
     auth.signin({
       connection: 'Username-Password-Authentication',
       username: $scope.user,
       password: $scope.pass,
       scope: 'openid profile'
+    }).then(onLoginSuccess, onLoginFailed)
+    .finally(function () {
+      $scope.loading = false;
     });
   };
 
   $scope.doGoogleAuthWithPopup = function () {
+    $scope.$parent.message = 'loading...';
+    $scope.loading = true;    
+
     auth.signin({
       popup: true,
-      connection: 'google-oauth2',
-      scope: 'openid profile'
+      connection: 'google-oauth2'
+    }).then(onLoginSuccess, onLoginFailed)
+    .finally(function () {
+      $scope.loading = false;
     });
   };
 
-  $scope.doGoogleAuthWithRedirect = function () {
-    return auth.signin({
-      connection: 'google-oauth2',
-      scope: 'openid profile'
-    });
-  };
 });
 
 myApp.controller('LogoutCtrl', function (auth, $scope, $location) {
   auth.signout();
+  $scope.$parent.message = '';
   $location.path('/login');
 });