+++ src/auth0-angular.js
@@ -2,21 +2,6 @@
 
   var util = angular.module('util', []);
 
-  util.factory('$safeApply', function safeApplyFactory($rootScope, $exceptionHandler) {
-    return function safeApply(scope, expr) {
-      scope = scope || $rootScope;
-      if (['$apply', '$digest'].indexOf(scope.$root.$$phase) !== -1) {
-        try {
-          return scope.$eval(expr);
-        } catch (e) {
-          $exceptionHandler(e);
-        }
-      } else {
-        return scope.$apply(expr);
-      }
-    };
-  });
-
   //this is used to parse the profile
   util.value('urlBase64Decode', function (str) {
     var output = str.replace('-', '+').replace('_', '/');
@@ -43,18 +28,21 @@
 
   auth0.constant('AUTH_EVENTS', AUTH_EVENTS);
 
-  function Auth0Wrapper(auth0Lib, $cookieStore, $rootScope, $safeApply, $q, urlBase64Decode) {
+  function Auth0Wrapper(auth0Lib, $cookieStore, $rootScope, $q, urlBase64Decode, $timeout) {
     this.auth0Lib     = auth0Lib;
     this.$cookieStore = $cookieStore;
     this.$rootScope   = $rootScope;
-    this.$safeApply   = $safeApply;
     this.$q           = $q;
+    this.$timeout     = $timeout;
 
     this.urlBase64Decode = urlBase64Decode;
 
     this.delegatedTokens = {};
 
     this.profile      = {};
+
+    this._loaded      = $q.defer();
+    this.loaded       = this._loaded.promise;
   }
 
   Auth0Wrapper.prototype = {};
@@ -168,38 +156,48 @@
       obj = obj.getClient();
     }
 
-    obj.getDelegationToken(clientID, this.idToken, options, this._wrapCallback(function (err, delegationResult) {
-      if (err) {
-        return deferred.reject(err);
-      }
+    obj.getDelegationToken(clientID, this.idToken, options, function (err, delegationResult) {
+      that.$timeout(function () {
+        if (err) {
+          return deferred.reject(err);
+        }
 
-      that.delegatedTokens[clientID] = delegationResult.id_token;
+        that.delegatedTokens[clientID] = delegationResult.id_token;
 
-      return deferred.resolve(delegationResult.id_token);
-    }));
+        return deferred.resolve(delegationResult.id_token);
+      });
+    });
 
     return deferred.promise;
   };
 
   Auth0Wrapper.prototype.signin = function (options) {
     options = options || {};
+    var that = this;
+    var $q = that.$q;
+    var defer = $q.defer();
 
     var callback = function(err, profile, id_token, access_token, state) {
-      if (err) {
-        that.$rootScope.$broadcast(AUTH_EVENTS.loginFailed, err);
-        return;
-      }
+      that.$timeout(function () {
+        if (err) {
+          that.$rootScope.$broadcast(AUTH_EVENTS.loginFailed, err);
+          defer.reject(err);
+          return;
+        }
 
-      that._serialize(id_token, access_token, state);
-      that._deserialize();
+        that._serialize(id_token, access_token, state);
+        that._deserialize();
 
-      that.getProfile(id_token)
+        that.getProfile(id_token)
         .then(function () {
           that.$rootScope.$broadcast(AUTH_EVENTS.loginSuccess, that.profile);
+          defer.resolve(that.profile);
+        }, function (err) {
+          that.$rootScope.$broadcast(AUTH_EVENTS.loginFailed, err);
+          defer.reject(err);
         });
+      });
     };
-
-    var that = this;
 
     // In Auth0 widget the callback to signin is executed when the widget ends
     // loading. In that case, we should not broadcast any event.
@@ -208,19 +206,13 @@
     }
 
     that.auth0Lib.signin(options, callback);
+    return defer.promise;
   };
 
   Auth0Wrapper.prototype.signout = function () {
     this._serialize(undefined, undefined, undefined);
     this._deserialize();
     this.$rootScope.$broadcast(AUTH_EVENTS.logout);
-  };
-
-  Auth0Wrapper.prototype._wrapCallback = function (callback) {
-    var that = this;
-    return function () {
-      return that.$safeApply(undefined, callback.apply(null, arguments));
-    };
   };
 
   Auth0Wrapper.prototype.parseHash = function (locationHash) {
@@ -231,23 +223,25 @@
     var deferred = this.$q.defer();
     var that = this;
 
-    var wrappedCallback = this._wrapCallback(function (err, profile) {
-      if (err) {
-        return deferred.reject(err);
-      }
+    var wrappedCallback = function (err, profile) {
+      that.$timeout(function () {
+        if (err) {
+          return deferred.reject(err);
+        }
 
-      // Cleanup old keys
-      Object.keys(that.profile).forEach(function (key) {
-        delete that.profile[key];
+        // Cleanup old keys
+        Object.keys(that.profile).forEach(function (key) {
+          delete that.profile[key];
+        });
+
+        // Add new keys
+        Object.keys(profile).forEach(function (key) {
+          that.profile[key] = profile[key];
+        });
+
+        deferred.resolve(that.profile);
       });
-
-      // Add new keys
-      Object.keys(profile).forEach(function (key) {
-        that.profile[key] = profile[key];
-      });
-
-      deferred.resolve(that.profile);
-    });
+    };
 
     this.auth0Lib.getProfile(token, wrappedCallback);
 
@@ -296,7 +290,7 @@
       $provide.value('auth0Lib', auth0Lib);
     };
 
-    this.$get = function ($cookieStore, $rootScope, $safeApply, $q, $injector, urlBase64Decode) {
+    this.$get = function ($cookieStore, $rootScope, $q, $injector, urlBase64Decode, $timeout) {
       // We inject auth0Lib manually in order to throw a friendly error
       var auth0Lib = $injector.get('auth0Lib');
       if (!auth0Lib) {
@@ -304,7 +298,15 @@
       }
 
       if (!auth0Wrapper) {
-        auth0Wrapper = new Auth0Wrapper(auth0Lib, $cookieStore, $rootScope, $safeApply, $q, urlBase64Decode);
+        auth0Wrapper = new Auth0Wrapper(auth0Lib, $cookieStore, $rootScope, $q, urlBase64Decode, $timeout);
+        if (!$injector.has('parseHash')) {
+          auth0Wrapper._deserialize();
+          auth0Wrapper.getProfile(auth0Wrapper.idToken)
+          .finally(function () {
+            auth0Wrapper._loaded.resolve();
+            $rootScope.$broadcast(AUTH_EVENTS.redirectEnded);
+          });
+        }
       }
 
       return auth0Wrapper;
@@ -312,29 +314,50 @@
   });
 
   var auth0Redirect = angular.module('auth0-redirect', ['auth0']);
+
+  auth0Redirect.factory('parseHash', function (auth, $rootScope, $window) {
     return function () {
       var result = auth.parseHash($window.location.hash);
 
-      if (result && result.id_token) {
-        // this is only used when using redirect mode
-        auth.getProfile(result.id_token).then(function () {
-          auth._serialize(result.id_token, result.access_token, result.state);
-          // this will rehydrate the "auth" object with the profile stored in $cookieStore
-          auth._deserialize();
-
-          $rootScope.$broadcast(AUTH_EVENTS.loginSuccess, auth.profile);
-        }, function (err) {
-          // this will rehydrate the "auth" object with the profile stored in $cookieStore
-          auth._deserialize();
-
-          $rootScope.$broadcast(AUTH_EVENTS.loginFailed, err);
-        });
-      } else {
+      function onAuthSuccess () {
+        auth._serialize(result.id_token, result.access_token, result.state);
+        // this will rehydrate the "auth" object with the profile stored in $cookieStore
         auth._deserialize();
+
+        $rootScope.$broadcast(AUTH_EVENTS.loginSuccess, auth.profile);
+      }
+
+      function onAuthFail(err) {
+        // this will rehydrate the "auth" object with the profile stored in $cookieStore
+        auth._deserialize();
+
+        $rootScope.$broadcast(AUTH_EVENTS.loginFailed, err);
+      }
+
+      function onRedirectEnded() {
+        auth._loaded.resolve();
+        $rootScope.$broadcast(AUTH_EVENTS.redirectEnded);
+      }
+
+
+      // this will rehydrate the "auth" object with the profile stored in $cookieStore
+      auth._deserialize();
+
+      // already logged in
+      if (auth.idToken) {
         auth.getProfile(auth.idToken)
-          .finally(function () {
-            $rootScope.$broadcast(AUTH_EVENTS.redirectEnded);
-          });
+          .then(onAuthSuccess, onAuthFail)
+          .finally(onRedirectEnded);
+      
+      // callback URL
+      } else if (result && result.id_token) {
+        auth.getProfile(result.id_token)
+        .then(onAuthSuccess, onAuthFail)
+        .finally(onRedirectEnded);
+
+      // page reloaded, not logged in
+      } else {
+        onRedirectEnded();
       }
     };
   });
@@ -343,9 +366,15 @@
 
   var authInterceptorModule = angular.module('authInterceptor', ['auth0']);
 
-  authInterceptorModule.factory('authInterceptor', function (auth, $rootScope, $q, AUTH_EVENTS) {
+  authInterceptorModule.factory('authInterceptor', function ($rootScope, $q, AUTH_EVENTS, $injector) {
     return {
       request: function (config) {
+        // When using auth dependency is never loading, we need to do this manually
+        // This issue should be related with: https://github.com/angular/angular.js/issues/2367
+        if (!$injector.has('auth')) {
+          return config;
+        }
+        var auth = $injector.get('auth');
         config.headers = config.headers || {};
         if (auth.idToken) {
           config.headers.Authorization = 'Bearer '+ auth.idToken;